Aktuell?

+++ Client/ImageManager.cs
@@ -13,8 +13,7 @@ namespace Client
         public ImageManager(string hostname, int port) : base(hostname, port)
         {
         }
-
-        [DebuggerStepThrough]
+        
         public byte[] GetImage()
         {
             Stream.WriteByte(Global.RequestImage);
@@ -25,21 +24,24 @@ namespace Client
 
             while (true)
             {
-                var buffer = Stream.ReadByte();
-                if (buffer != -1) byteList.Add((byte)buffer);
-
-                // If the last 4 bytes are 0xF, 0x0, 0xf, 0x0, end transmission and convert list to array.
-                var length = byteList.Count;
-                if ((length > 3) &&
-                    (byteList[length - 1] == 0xAF) &&
-                    (byteList[length - 2] == 0xFA) &&
-                    (byteList[length - 3] == 0xAF) &&
-                    (byteList[length - 4] == 0xFA))
+                if (Stream.DataAvailable)
                 {
-                    // Removes the 4 bytes that are used to indicate end of stream.
-                    byteList.RemoveRange(length - 4, 4);
-                    image = byteList.ToArray();
-                    break;
+                    var buffer = Stream.ReadByte();
+                    if (buffer != -1) byteList.Add((byte)buffer);
+
+                    // If the last 4 bytes are 0xF, 0x0, 0xf, 0x0, end transmission and convert list to array.
+                    var length = byteList.Count;
+                    if ((length > 3) &&
+                        (byteList[length - 1] == 0xAF) &&
+                        (byteList[length - 2] == 0xFA) &&
+                        (byteList[length - 3] == 0xAF) &&
+                        (byteList[length - 4] == 0xFA))
+                    {
+                        // Removes the 4 bytes that are used to indicate end of stream.
+                        byteList.RemoveRange(length - 4, 4);
+                        image = byteList.ToArray();
+                        break;
+                    }
                 }
             }
             return image;
